IMPALA-1330: Fix column types for SHOW {table, partition} STATS
Because we add 'total' to the last row in SHOW PARTITIONS, we set the
partition key columns to be string. At least, that's what the comment
said, but we didn't do that in fact.

This patch also corrects the column type for max width, which should be INT.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4678
Tested-by: jenkins

+++ fe/src/main/java/com/cloudera/impala/service/Frontend.java
@@ -606,7 +606,7 @@ public class Frontend {
     resultSchema.addToColumns(
         new TColumn("#Distinct Values", Type.BIGINT.toThrift()));
     resultSchema.addToColumns(new TColumn("#Nulls", Type.BIGINT.toThrift()));
-    resultSchema.addToColumns(new TColumn("Max Size", Type.DOUBLE.toThrift()));
+    resultSchema.addToColumns(new TColumn("Max Size", Type.INT.toThrift()));
     resultSchema.addToColumns(new TColumn("Avg Size", Type.DOUBLE.toThrift()));
 
     for (Column c: table.getColumnsInHiveOrder()) {
